reduce long methods
